name_value/ptree.h: added todo for guessing ini format
Currently, ini will not work since path-value parser treats ini sections and comments as valid path-value records rather than failing.
Possible solution is to implement stricter checks in path-value parser so that it fails on ini files with sections and/or comments.
Disabled tests that illustrate the problem are added to serialise_guess_test.cpp.