io-bandwidth: correctly calculate bandwidth after a pause by consumer
If the consumer of stdout from io-bandwidth stops consuming the buffer
will fill and then io-bandwidth will stop as well. If the consumer then
restarts the buffer will empty but next_bucket and next_update will be a
long way behind actual time. Until they catch up (normally quick but if
the pause was long the catch-up will take a while as well) the
calculations will be wildly inaccurate.

Fix this by just resetting to the current time.