csv-time-join: remove --no-discard option
it was a work-around for the bug where points before or after the bounding
data would be discarded as unable to be consistently timestamped. That is
now fixed and those points are kept (and consistently handled) so this
option is no longer required. (it's behaviour was quite inconsistent in
any case).